fix vim install.sh

+++ vimrc/install.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
-mv ~/.vim ~/.vim.bak
-mv ~/.vimrc ~/.vimrc.bak
+# back
+if [ -f ~/.vim ]; then
+    mv ~/.vim ~/.vim.bak
+fi
 
+if [ -f ~/.vimrc ]; then
+    mv ~/.vimrc ~/.vimrc.bak
+fi
+
+# install
 cp -r vim ~/.vim
 cp vimrc ~/.vimrc
